Silence compiler warning.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@21327 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseMerge/BaseView.cpp
@@ -3434,7 +3434,7 @@ void CBaseView::PasteText()
     CString sLine = sClipboardText.Mid(nStart, sClipboardText.GetLength() - nStart);
     lines.push_back(sLine);
 
-    int nLinesToPaste = lines.size();
+    int nLinesToPaste = (int)lines.size();
     if (nLinesToPaste > 1)
     {
         CString sLine = GetViewLineChars(nViewLine);
